week2/Day1/requirements.txt
# Day 1 Requirements
openai>=1.12.0
python-dotenv>=1.0.0
